functions/processor.py
import os
import json
import logging
import mimetypes
from google.cloud import firestore, storage
import requests
from typing import Dict, Any, Optional

logger = logging.getLogger(__name__)

class MediaProcessor:
    def __init__(self):
        self.firestore_client = firestore.Client()
        self.storage_client = storage.Client()
        self.backend_callback_url = os.getenv("BACKEND_CALLBACK_URL", "http://localhost:8000/api/v1/media-callback")
        self.max_cost = float(os.getenv("MAX_PROCESSING_COST_USD", "10.0"))

    def process_media_event(self, bucket_name: str, file_name: str, event_id: str) -> Dict[str, Any]:
        """
        Main orchestrator for processing uploaded media files.
        """
        try:
            # Check for idempotency
            if self._is_event_already_processed(event_id):
                logger.info(f"Event {event_id} already processed, skipping")
                return {"success": True, "message": "Already processed"}

            # Mark event as processing
            self._mark_event_processing(event_id)

            # Detect media type
            media_type = self._detect_media_type(file_name)
            if not media_type:
                raise ValueError(f"Unsupported media type for file: {file_name}")

            # Download file temporarily (for local processing)
            local_path = self._download_file(bucket_name, file_name)

            # Process based on media type
            processed_results = self._process_media(local_path, media_type, bucket_name, file_name)

            # Generate thumbnail if applicable
            thumbnail_path = self._generate_thumbnail(local_path, media_type, bucket_name, file_name)

            # Update processed results
            processed_results.thumbnail_path = thumbnail_path

            # Send callback to backend
            callback_success = self._send_backend_callback(event_id, f"gs://{bucket_name}/{file_name}", processed_results)

            if callback_success:
                # Mark event as completed
                self._mark_event_completed(event_id)
                return {"success": True, "processed_results": processed_results.dict()}
            else:
                raise Exception("Backend callback failed")

        except Exception as e:
            logger.error(f"Error processing media event: {str(e)}")
            self._mark_event_failed(event_id, str(e))
            return {"success": False, "error": str(e)}

    def _is_event_already_processed(self, event_id: str) -> bool:
        """Check if event has already been processed."""
        doc_ref = self.firestore_client.collection("processing_events").document(event_id)
        doc = doc_ref.get()
        if doc.exists:
            status = doc.to_dict().get("status")
            return status in ["completed", "processing"]
        return False

    def _mark_event_processing(self, event_id: str):
        """Mark event as currently processing."""
        doc_ref = self.firestore_client.collection("processing_events").document(event_id)
        doc_ref.set({
            "status": "processing",
            "started_at": firestore.SERVER_TIMESTAMP
        })

    def _mark_event_completed(self, event_id: str):
        """Mark event as completed."""
        doc_ref = self.firestore_client.collection("processing_events").document(event_id)
        doc_ref.update({
            "status": "completed",
            "completed_at": firestore.SERVER_TIMESTAMP
        })

    def _mark_event_failed(self, event_id: str, error: str):
        """Mark event as failed."""
        doc_ref = self.firestore_client.collection("processing_events").document(event_id)
        doc_ref.update({
            "status": "failed",
            "error": error,
            "failed_at": firestore.SERVER_TIMESTAMP
        })

    def _detect_media_type(self, file_name: str) -> Optional[str]:
        """Detect media type from file extension."""
        mime_type, _ = mimetypes.guess_type(file_name)
        if mime_type:
            if mime_type.startswith("image/"):
                return "image"
            elif mime_type.startswith("video/"):
                return "video"
            elif mime_type.startswith("audio/"):
                return "audio"
        return None

    def _download_file(self, bucket_name: str, file_name: str) -> str:
        """Download file from GCS to temporary local path."""
        bucket = self.storage_client.bucket(bucket_name)
        blob = bucket.blob(file_name)
        local_path = f"/tmp/{file_name.split('/')[-1]}"
        blob.download_to_filename(local_path)
        return local_path

    def _process_media(self, local_path: str, media_type: str, bucket_name: str, file_name: str) -> Dict[str, Any]:
        """Process media based on type using Vertex AI and other services."""
        # TODO: Implement actual Vertex AI calls
        # For now, return mock processed results

        processed_results = {
            "media_type": media_type,
            "create_product": False  # Default to False, can be set based on analysis
        }

        if media_type == "image":
            # Mock vision analysis
            processed_results.update({
                "vision_analysis": {"objects": ["placeholder"], "confidence": 0.9},
                "generated_description": "A placeholder image description",
                "embeddings": [0.1, 0.2, 0.3]  # Mock embeddings
            })
        elif media_type == "video":
            # Mock video analysis
            processed_results.update({
                "transcription": "Placeholder video transcription",
                "vision_analysis": {"scenes": ["placeholder_scene"]},
                "generated_description": "A placeholder video description"
            })
        elif media_type == "audio":
            # Mock audio transcription
            processed_results.update({
                "transcription": "Placeholder audio transcription"
            })

        # TODO: Add logic to determine if product should be created
        # For example, based on content analysis or user preferences

        return processed_results

    def _generate_thumbnail(self, local_path: str, media_type: str, bucket_name: str, file_name: str) -> Optional[str]:
        """Generate thumbnail for media file."""
        # TODO: Implement actual thumbnail generation
        # For now, return None or mock path
        return None

    def _send_backend_callback(self, event_id: str, gcs_path: str, processed_results: Dict[str, Any]) -> bool:
        """Send processed results to backend callback endpoint."""
        try:
            payload = {
                "event_id": event_id,
                "gcs_path": gcs_path,
                "processed_results": processed_results
            }

            # TODO: Add authentication token if required
            response = requests.post(self.backend_callback_url, json=payload, timeout=30)

            if response.status_code == 200:
                logger.info(f"Successfully sent callback for event {event_id}")
                return True
            else:
                logger.error(f"Backend callback failed with status {response.status_code}: {response.text}")
                return False

        except Exception as e:
            logger.error(f"Error sending backend callback: {str(e)}")
            return False

def process_media_event(bucket_name: str, file_name: str, event_id: str) -> Dict[str, Any]:
    """Entry point for processing media events."""
    processor = MediaProcessor()
    return processor.process_media_event(bucket_name, file_name, event_id)
