functions/requirements.txt
functions-framework>=3.5.0
firebase-functions==0.1.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.10.0
vertexai==1.43.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
firebase-admin==6.2.0
